refactor: use uint32 for MateLen

+++ source/engine/user-engine/mate_len.hpp
@@ -121,7 +121,7 @@ class MateLenImpl : DefineNotEqualByEqual<MateLenImpl<T>>, DefineComparisonOpera
  *
  * 範囲外の初期値がほしい場合は、`kMinus1MateLen` や `kDepthMaxPlus1MateLen` を用いる。
  */
-using MateLen = detail::MateLenImpl<std::uint16_t>;
+using MateLen = detail::MateLenImpl<std::uint32_t>;
 
 /**
  * @brief 詰み／不詰手数（`MateLen` の16ビット版）
